[batch] Sort jobs by job id in the UI (#14585)
These queries are responsible for getting the paginated list of jobs in
the batch UI. I believe this missing order by went unnoticed because the
queries were naturally getting returned in order based on the jobs
primary key, but we've recently seen a couple page loads of jumbled or
missing jobs. Seems like something around the query stats changed and
the database adjusted its plan slightly. Regardless, we want this order
by explicit.