[batch] Adjust Batch UI to have less dead space (#14587)
The original UI redesign was intended to make it easier to navigate and
focus on important information. I received very positive feedback on
some aspects like this, for example having the different job steps in
separate tabs. There was also feedback though that there was a lot of
dead space. This is an attempt to find a good middle ground where the UI
is not overwhelmingly cluttered but is more compact and makes better use
of screen real estate