[query] Chunk spark backend execution (#14590)
We seem to be running into spark/yarn scheduling limitations causing
stages to often fail with a large number of partitions. Here, we
implement a very simple chunking strategy to run spark jobs with a
limited number of partitions at a time.

The maximum parallelism is controlled by a new
`spark_max_stage_parallelism` feature flag, which defaults to MAXINT
until we can figure out a good default.

Also, this change corrects a small error in logic for partition indices
for call caching. The `resultHandler` argument of [`runJob`] is called
with the job's partition index, not the index of the partition within
the RDD. So we need to index into the `partitions` sequence when
populating the results buffer.

CHANGELOG: Add 'spark_max_stage_parallelism' flag to allow users to run
pipelines with a large number of partitions in chunks. By default, hail
still attempts to run all partitions in a stage at once.

[`runJob`]:
https://spark.apache.org/docs/latest/api/scala/org/apache/spark/SparkContext.html#runJob[T,U](rdd:org.apache.spark.rdd.RDD[T],func:(org.apache.spark.TaskContext,Iterator[T])=%3EU,partitions:Seq[Int],resultHandler:(Int,U)=%3EUnit)(implicitevidence$11:scala.reflect.ClassTag[U]):Unit