[batch] Fix incomplete API responses in idempotency-handling branches (#14589)
`'/api/v1alpha/batches/create-fast'` is a Batch API endpoint that allows
you to create a batch, submit jobs, and start them running all in one
HTTP request. It makes a few transactions within the endpoint and must
be idempotent. One of the failure modes it must watch for is when a
Batch Update to add some new jobs and/or job groups to a batch has
already successfully been committed, but the request is retried for X
reason. In that scenario, we want to just return a 200 to the client. We
do this already, but there was an oversight when creating job groups
where those branches did not include the full response payload, which
results in intermittent errors on the client when it can't retrieve
`resp['last_job_group_id']`. This fixes that so we always return the
same JSON.