[query] Fix deserialization of IR function parameter names in QoB (#14579)
I believe #14547 introduced a bug that broke IR function deserialization
in QoB by changing `value_parameter_names` from `Array[String]` to
`Array[Name]`. This fixes the issue by pushing the introduction of the
`Name` wrapper object to after deserialization. Another option is to
incorporate the `{ str: String }` structure of `Name` into the python ->
scala payload, but I'm not sure I see a use case for that and we can
always do that later (there is no issue of backwards compatibility with
this communication between python and scala).

My main concern here is that clearly this isn't tested. I'd appreciate
guidance on the current advised practice for testing this behavior.