[auth] Make UI consistent with new Batch design (#14586)
This aligns the auth styling with the new batch UI styling. Nothing has
really changed about the structure of the pages. This one should be
considerably easier than the batch overhaul, just three pages `/user`,
`/users` and `/roles` (I think roles were added to auth but never
fleshed out, we never use this page. But that's for another PR..).

I manually checked that I didn't break any of the Batch pages after
moving the utils file into web_common, but a sanity check on that would
be great.