[batch] Fix accidental mutation in GoogleComputeClient.list that broke disk cleanup (#14614)
Workers can dynamically attach disks to themselves to accommodate jobs
that request storage that the VM cannot accommodate. In certain
circumstances like preemption, the VM can disappear before it is able to
delete its own disks, so the Batch Driver scans for disks that are no
longer attached (orphaned) and deletes them. It looks like our disk
cleanup loop was broken due to inadvertent mutation that leads to an
assertion error if the same `params` argument is used in multiple
invocations of `GoogleComputeClient.list`. This is preventing orphaned
disks from being deleted, costing us money. Additional details are in
#14613.


Fixes #14613