[ci] Add operator knob to adjust rate limits (#14609)
This adds a control to CI where an operator can adjust the rate limit on
a particular service in a particular namespace. CI stores and propagates
that information when it generates the Envoy configs for `gateway` and
`internal-gateway`. This enables operators to shield an overwhelmed
batch driver by decreasing its rate limit without needing to push a code
change. Increasing the rate limit can increase throughput if workers are
being rate limited but the batch driver is not fully utilized.