[gateway] Fix top level domain for website (#14617)
#14609 broke the routing for https://hail.is. The `domains` variable
here indicates to `gateway` which domains in incoming requests should be
routed to the given service. Since #14609 changed the `service`
parameter from `str` to `Service`, it silently broke this branch.

This wasn't covered by the envoy config generation tests because we
didn't have a `www` service in the test configuration. I've added it so
that this branch is covered.

Fixes #14616