[dataproc] adds project to args for submit
Currently, if you try to pass `--project` to `hailctl dataproc submit`,
it still tries to use the project configured via `gcloud config`, and
passes through the value specified for `--project` in the
`pass_through_args`.

This change allows the user to specify which project their Dataproc
cluster is in as part of the command.

An example command where the project is ignored on `main` but used
correctly on this branch is:

```bash
hailctl dataproc submit test.py --region us-central1 --project broad-ctsa
```