[query] Remove jackson string length restriction in Spark/Local backends (#14651)
As mentioned in #14580, IR can get quite big, especially as it can
contain an arbitrary amount of encoded literals from the user's python
session.

Tested manually, by making a very very large literal, running a pipeline
with it on 0.2.132, observing the failure seen in #14650, then running
the same pipeline with this change, and it succeeds as normal.

Resolves #14650