[ci/azure] installs dependencies for azure-cli image (#14677)
The `delete_azure_batch_instances` step is failing on various PRs with
the error `jq: command not found`. This appears to be because we do not
pin the version for the `mcr.microsoft.com/azure-cli` image, and while
that image was previously based on the Alpine image, [now it is based on
the Azure Linux
image](https://learn.microsoft.com/en-us/cli/azure/run-azure-cli-docker),
and does not appear to have `jq` (or `kubectl`) preinstalled on it.

This change updates the commands run in the `azure-cli` container for
this CI step to install `jq` and `kubectl` via `curl` before running the
relevant commands. The `curl` commands were tested locally by running
`docker run -it mcr.microsoft.com/azure-cli` and trying them out in the
image's shell.

This change also adds the installation commands in the other place where
this image is used (when cleaning up from `buildImage2` jobs that are
run in Azure).