[query] Set QUAL to missing when it's "."
When we switched from using htsjdk to do most of our VCF parsing to
using our own parser, we kept the behavior of setting missing QUAL to
a sentinel value. This isn't necessary (or correct) for hail to do. The
type of QUAL is optional float64 and we should preserve the missing
value.

Technically speaking this represents a backwards compatibility change.
However, I think this is a bug in my attempt in the past to keep
bug-for-bug compatibility with htsjdk rather than parsing the value into
what people would expect.