[vep] Use system package manager for perl libraries (#14713)
### Change Description

The changes in #14708 caused our vep images to be rebuilt, breaking
them. The solution here is the one that we should have done from the
beginning, install dependencies through the system package manager. In
the interest of my sanity and effort, I only added the libraries that
were being problematic from CPAN to the list of system installed ones.

### Security Assessment
- [x] This change has no security impact