[query/vds] Add LEN field to VDS  (#14675)
This is the beginning of a series of changes to support export of VDS to
VCF 4.5, the version of VCF that contains the standardized form of our
work that culminated in SVCR/VDS.

Reference blocks were standardized with a LEN rather than an END. So,
now, by default, add LEN to all VDS reads and drop END in favor of LEN
on all VDS writes. Our optimizer will be able to take care of pruning
away the dead field in pipelines that don't use it.

We make sure that all VDS creation (other than the combiner), such as
read_vds and from_merged_representation, contains both LEN and END
preserving user code that depends on the presence of the END field.

Furthermore, this change contains necessary combiner updates to prefer
LEN over END, and to use LEN in the combiner itself.

---------

Co-authored-by: Patrick Schultz <pschultz@broadinstitute.org>