[batch] Handle duplicate resource names when refreshing prices
It appears that GCP has introduced products with identical names and
versions, so inserting into the resources table is failing with primary
key uniqueness errors. This is not a good solution, but at current, this
issue is blocking development and any solution must be implemented. If
the `resource_name` is duplicated and the prices are the same for both
resources, then we're fine. Otherwise, this change will throw an error.
Thus, if resource names ever diverge in price while having the same
name, we'll need to find a more robust solution.