[batch] During rate refresh, update SKU if it is NULL
Right now, our `latest_product_version` table has a partially filled sku
column. This is because we added sku and didn't back calculate the old
skus, and we haven't assigned skus to 'our' products (ip fees, service
fee). I'd like to fix that, so that most of our 'resources' can be
tracked by sku rather than some of the more fragile name/version strings
that has lead to conflicts with duplicates in GCP.