Update Homebrew Java installation docs for OSX (#14729)
## Change Description

When going through the [documentation for installing Hail on
OSX](https://hail.is/docs/0.2/install/macosx.html), I noticed that the
syntax for installing Java via Homebrew was out of date.

This PR updates the documentation to use the latest syntax for Homebrew.
It also updates the command to install version 11 of Temurin instead of
version 8.

## Security Assessment

- This change has no security impact

### Impact Description

This change updates documentation only and has no immediate end user
impact.

This change updates the recommended version of Temurin to a newer
version (11 vs. 8). It is reasonable to assume that the newer version is
at least as secure as previous versions. So, there also should be no
negative security impact on future users of this documentation.