[batch] Openapi doc for batch service (#14777)
## Change Description

Adds two endpoints in the `batch` service hosting an openapi
specification and a swagger page.

**Note to reviewers:** The API specification documentat itself is nice
to get right, but relatively low risk. I'd prioritize time making sure
the framework to host it is correct, and the openapi setup sections
before scanning through the details of every individual API call.

## Security Assessment

Delete all except the correct answer:
- This change has a high security impact
  - [x] Required: The impact has been assessed and approved by appsec

### Impact Description

We are adding two new endpoints in `batch`. Mitigating factors:

- There are no user parameters for these endpoints
- The pages being hosted are mostly static, apart from the "base url"
which the API documentation and swagger page targets, which is generated
in the backend.

(Reviewers: please confirm the security impact before approving)