move ir node defs into subpackage of ir (#14783)
## Change Description

Move the definitions of the IR nodes from package `ir` into a subpackage
`ir.defs`. I originally did this to prepare for codegenerating the node
definitions. While I don't think it was necessary in the end, I still
think this is a beneficial change, as the `ir` package has become a big
pile of different functionality, and it would be good to start breaking
it up.

## Security Assessment

- This change has no security impact

### Impact Description

Non-functional change, only changes the java package paths of some
classes.

(Reviewers: please confirm the security impact before approving)