[query] Expose references via `ExecuteContext` (#14686)
This change is split out from a larger refactoring effort on the various
Backend
implementations. The goals of this effort are to provide query-level 
configuration to the backend that's currently tied to the lifetime of a
backend,
 reduce code duplication and reduce state duplication.

In this change, I'm restoring references to the execute context [1] and 
decoupling them from the backend. In a future change, they'll be lifted
out of
the backend implementations altogether. This is to reduce the surface
area of
the Backend interface to the details that are actually different.

Both the Local and Spark backend have state that's manipulated from
python via
various py methods. These pollute the Backend interface [2] and so have
been
extracted into the trait Py4JBackendExtensions. In future changes, this
will
become a facade that owns state set in python.

Notes
[1] "Restoring" old behaviour I foolishly removed in fe5ed32
[2] "Pollute" in that they obfuscate what's different about backend
query plan
     and execution