[query] Preserve field order in StructExpression.rename (#14807)
Either set construction from an iterator or set difference is not order
preserving. If we iterate through the struct fields in order, then our
select inside of rename will preserve order for existing fields.

Brief description and justification of what this PR is doing.

## Security Assessment

Delete all except the correct answer:
- This change has no security impact

### Impact Description
Query only. Simply causes a single iteration to happen in a different
order.