[query] StagedIndexReader now requires an index spec
For legacy reasons, the data contained by an index leaf node includes
a field called `offset` and another called `annotation`. The `offset` is
always an int64, and `annotation` contains arbitrary data. When writing
a matrix table, we use a field in `annotation` called `entries_offset`.
In the index spec, we make a note of if there is a different offset
than `offset`.

This means, that if we have an index spec, we can't assume that the
offset field is always `offset`, and so, add a helper method that takes
a leaf node, and give us the offset the index spec is expecting.