[query] Remove BlockMatrix persist from Backend interface (#14690)
This change is split out from a larger refactoring effort on the various
Backend
implementations. The goals of this effort are to provide query-level
configuration to the backend that's currently tied to the lifetime of a
backend,
reduce code duplication and reduce state duplication.

In this change, I'm removing blockmatrix persist/unpersist from the
`Backend`
interface by adding `BlockMatrixCache: mutable.Map[String, BlockMatrix]`
to
`ExecuteContext`. The various reader/writer implementations simply fetch
the
block matrix from this cache. For the spark backend, this is backed by a
cache
whose lifetime is tied to the spark backend. Since block matrices are
not
supported in the local and service backends, the cache is an empty map.

Note that block matrix persist is broken in python (#14689)