[auth] Allow re-login after session timeout (#14818)
## Change Description

Currently: when a user's session times out or becomes invalid for any
reason, they are redirected to a login page which (because of the
existing but stale session cookie) just returns 401: Unauthorized.

With this change: we allow users with invalid or stale session cookies
to access the user page and re-login

## Security Assessment

- This change has a medium security impact

### Impact Description

Allow access to pages that previously would have failed when a user's
session was invalid.
However: 
- The pages in question are accessible for unauthorized users
- We treat the page access as though the user were unauthorized

So this is essentially equivalent to the user clearing out their cookie
cache before attempting to access the page.