[batch] Add support for n1 machines with T4 GPUs in GCP (#14844)
## Change Description

Adds support for n1 machines with T4 GPUs in GCP. Fixes [a previous PR's
attempt at adding this
functionality](https://github.com/hail-is/hail/pull/14821) (note:
regarding billing, this was actually addressed properly in that previous
PR, funny enough). Convention for specifying a n1+T4 machine in batch
for a job j is `j._machine_type = '{name of n1 family
machine}+nvidia-tesla-t4+{num. GPUs}'`. Currently, this PR only supports
`{num. GPUs} = 1`.

As part of this PR, the Nvidia driver version has been updated to LTS
version 535.183.01 (from prev. 530.30.02). This is because 535.183.01 is
a solid [long-term support driver
version](https://docs.nvidia.com/datacenter/tesla/drivers/index.html#cuda-drivers),
and what [Google recommends as a minimum required version for using its
T4
GPUs](https://cloud.google.com/compute/docs/gpus/install-drivers-gpu).
This newer version has been tested with our current GPU machines (e.g.,
G2 w/ L4) and doesn't break anything.

Output of running `nvidia-smi` in a job on an n1 machine w/ T4 GPU is
below, and confirms that the machines are i) accessible, and ii) can use
their GPUs, which was the point of this PR in the first place.

```
Sun Mar 30 20:17:21 2025       
+---------------------------------------------------------------------------------------+
| NVIDIA-SMI 535.183.01             Driver Version: 535.183.01   CUDA Version: 12.2     |
|-----------------------------------------+----------------------+----------------------+
| GPU  Name                 Persistence-M | Bus-Id        Disp.A | Volatile Uncorr. ECC |
| Fan  Temp   Perf          Pwr:Usage/Cap |         Memory-Usage | GPU-Util  Compute M. |
|                                         |                      |               MIG M. |
|=========================================+======================+======================|
|   0  Tesla T4                       Off | 00000000:00:04.0 Off |                    0 |
| N/A   46C    P0              28W /  70W |      2MiB / 15360MiB |      0%      Default |
|                                         |                      |                  N/A |
+-----------------------------------------+----------------------+----------------------+
                                                                                         
+---------------------------------------------------------------------------------------+
| Processes:                                                                            |
|  GPU   GI   CI        PID   Type   Process name                            GPU Memory |
|        ID   ID                                                             Usage      |
|=======================================================================================|
|  No running processes found                                                           |
+---------------------------------------------------------------------------------------+
hello
```
### Updating and using the new worker image
Our Batch worker image is created/updated via
[gcp-create-worker-image.sh](https://github.com/hail-is/hail/blob/6e2393d2b018099f96f6c71b20543518276e6f9d/batch/gcp-create-worker-image.sh),
and as of this PR is on Version 17. Part of this shell script entails
running a startup script on our build image instance, specified to be
[build-batch-worker-image-startup-gcp](https://github.com/hail-is/hail/blob/6e2393d2b018099f96f6c71b20543518276e6f9d/batch/build-batch-worker-image-startup-gcp.sh),
where the Nvidia driver version is specified (and has been updated to be
535.183.01 in this PR). The `gcp-create-worker-image` script is run
manually, and was done so prior to the merging of this PR.

Once a worker image is created, [we can now point to as a source image
it when we want to create an
instance](https://github.com/hail-is/hail/blob/6e2393d2b018099f96f6c71b20543518276e6f9d/batch/batch/cloud/gcp/driver/create_instance.py#L120)
(updated accordingly to Version 17 in this PR). The `sourceImage` field
where we specify a worker image is written according to Google Cloud's
REST API documentation for creating VMs. For reference,
[here](https://cloud.google.com/compute/docs/gpus/create-gpu-vm-general-purpose#rest)
is the documentation for creating an N1 machine with or without a GPU,
and
[here](https://cloud.google.com/compute/docs/gpus/create-gpu-vm-accelerator-optimized#rest)
is the documentation for creating an A or G family machine.


## Security Assessment
- This change has a low security impact

### Impact Description
Adds additional machines available for running batch jobs, but does not
add any functionality of batch beyond what users are already able to do.

(Reviewers: please confirm the security impact before approving)

---------

Co-authored-by: Christopher Vittal <chris@vittal.dev>