[qob] Add ability to attach to existing batch (#14829)
The goal of this change is to be able to group QoB queries together to
allow for
easier visibility of how much a pipeline costs as well as make the
batches page
less cluttered so it's easier to tell what is running.

Adds `batch_id` parameter to `hl.init` that, when used with the
`'batch'` backend,
allows users to add query-on-batch jobs to an existing batch. Note that
racing two
or more QoB jobs concurrently in the same batch will likely fail.
Instead, users
should add dependences between query jobs so that they're scheduled
consecutively.

This change has low security impact as it only modifies client code:
users cannot
attach to batches that they do not already have access to.

---------

Co-authored-by: Edmund Higham <edhigham@gmail.com>