[compiler] move TableIR execute methods out of class (#14872)
## Change Description

Refactor the `TableIR` class to move the `execute` method out of the
class, into a free function. Besides just being a better organization,
this is also a step towards codegenerating the `TableIR` nodes.

Working through this, I had to stop myself from trying to rewrite some
`execute` methods to use the lowered implementation (as
`TableLeftJoinRightDistinct` has for some time). But this is a big
enough change, I wanted to keep it a pure refactoring. And when we start
replacing some RVD-based execute cases with lowering-based ones, we
should closely monitor the performance impact.

## Security Assessment

Delete all except the correct answer:

- This change has no security impact

### Impact Description

Non-functional change to the query compiler

(Reviewers: please confirm the security impact before approving)