[query] codegen all value ir node classes (#14771)
## Change Description

Generates the IR node class definitions from a declarative
specification, along with methods that replace the `Children` and `Copy`
functions.

The code generator is defined in a new mill module `ir-gen`. Before
compiling the main `hail` module, it will compile and execute the
`ir-gen` module, and include its output in the sources for `hail`.

Note that the structure of the code generator itself is something I want
to improve, and I have work in progress to do that.

Longer term, we might
- expand the code generator to generate printers and parsers (which
would automatically be kept in sync!)
- encode variable binding information and generate the `Binds.scala`
code
- generate generic IR traversal methods which decouple the details of
the IR data structure from the passes that operate on it, so that it's
easy to make changes to the data structure (e.g. only store an array of
children, not separate fields for each child, or potentially even move
to a mutable IR with parent pointers)
- use this to also generate the python IR classes and pretty printers,
which again would be automatically kept in sync

## Security Assessment

- This change has a low security impact

### Impact Description

Should be a non-functional change