[query/vds] Reorder LEN field to end of new ref entry struct
As of #14675, we are automatically adding a `LEN` field to VDS on read,
while also dropping the `END` field on write. For older VDS, this puts
the `LEN` field at the end of the reference entry struct. PR #14675 puts
`LEN` as the first item in the entry struct when creating new reference
data from a GVCF. This leads to an incompatibility when running the
combiner on a mixture of old and new datasets. The solution here can be
very simple. Just put `LEN` at the end of the entry struct on GVCF
import as well.