[query] Cache movie lens data for private use (#14893)
We often run into failures when the movie lens dataset's TLS certificate
expires. This is annoying. It's also annoying that we can't host this
data ourselves per the terms of service. However, there should be no
issues saving the data for internal use where other users cannot access
it.

If the `HAIL_USE_PRIVATE_LENS_DATA` environment variable is present and
nonempty, then we attempt to copy the data saved in GCS, otherwise, we
try the download as normal.

## Security Assessment
- This change has no security impact
### Impact Description
Query only.