[batch] Add default_region API endpoint (#14908)
## Change Description

Adds a new API endpoint to the Batch front end server to get the region
the Batch instance is running in.

This will be used in a future change for clients to be able to set the
region automatically to this default region.

## Security Assessment

- This change impacts the Hail Batch instance as deployed by Broad
Institute in GCP

### Impact Rating

- This change has a low security impact

### Impact Description

Addition of API endpoint that returns a single value that is just a
global parameter value of the system configuration.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec