[batch] Use the default region when a user doesn't specify a region (#14848)
## Change Description

BREAKING CHANGE: This changes the behavior of what the default regions
are set to from Any Region to the Batch Hosted Region.

This PR is intended to make the default behavior of the hailtop.batch
client use only the default region if no regions are specified. We need
to make this change because this is an obvious place for cost savings as
knowing to set the regions may not be conveyed to new people or others
may forget when getting a new laptop.

There are other ways to implement this change by using environment
variables instead of "Global Config" objects. I'm not particular on how
it's actually implemented.

## Security Assessment

- This change has a low security impact

### Impact Description

Added a new API endpoint to the Batch front end. It should only be
accessible to Authorized Users.

(Reviewers: please confirm the security impact before approving)

---------

Co-authored-by: Christopher Vittal <cvittal@broadinstitute.org>