[query] Serialize plink variants per partition (#14905)
PLINK bim records are not huge, but we were observing issues when
broadcasting very large numbers of variants leading to unexpected
failures in serialization.

We attempt to fix this by serializing each partition's variants to a
separate file that is decoded at the beginning of a pipeline.

## Security Assessment
- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP