[query] add methods for (multivariate) hypergeometric distribution (#14923)
## Change Description

Adds methods to sample from a hypergeometric and multivariate
hypergeometric distribution (`rand_hyper` and `rand_multi_hyper`), as
well as a function to evaluate the probability function of a
hypergeometric distribution (`phyper`). I chose a parameterization
consistent with
[scipy](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.hypergeom.html).
I believe the included doctests are sufficient to test these methods,
which simply expose library methods.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP